% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{besd_dictionary}
\alias{besd_dictionary}
\title{BeSD item dictionary}
\usage{
besd_dictionary()
}
\value{
A tibble with item metadata.
}
\description{
Built-in harmonised dictionary of core quantitative BeSD items. Questions and responses
align precisely with the BeSD quantitative toolkit see:
\url{https://www.who.int/publications/i/item/9789240049680} (accessed 24 Feb 2026)
}
