% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-besd.R
\name{besd_guess_topbox_levels}
\alias{besd_guess_topbox_levels}
\title{Heuristic for selecting "top-box" responses}
\usage{
besd_guess_topbox_levels(
  responses,
  prefer_regex = c("^Yes$", "Strongly agree|Completely agree", "Agree a lot|Agree",
    "Very likely|Definitely", "A lot|High")
)
}
\description{
Heuristic for selecting "top-box" responses
}
