% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{besd_palette_imms}
\alias{besd_palette_imms}
\title{IMMS palette}
\usage{
besd_palette_imms(
  n,
  reverse = FALSE,
  alpha = 1,
  anchors = c("#CC278D", "#926F97", "#4F8D9A")
)
}
\description{
IMMS palette
}
