% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd-summary.R
\name{besd_summary_demographics}
\alias{besd_summary_demographics}
\title{Extract demographic summaries}
\usage{
besd_summary_demographics(x)
}
\arguments{
\item{x}{A result from \code{summary(besd_data, include_demographics = TRUE)}.}
}
\value{
A tibble of demographic summaries, or \code{NULL} if not present.
}
\description{
When \code{summary(besd_data, include_demographics = TRUE)} is used, demographic
summaries are attached to the returned object as an attribute. This helper
retrieves that attribute.
}
