% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assert.R
\name{.assert_besd_summary_tbl}
\alias{.assert_besd_summary_tbl}
\title{Assert that an object is a usable BeSD summary table}
\usage{
.assert_besd_summary_tbl(x, fn = NULL, require_attrs = TRUE)
}
\description{
Assert that an object is a usable BeSD summary table
}
\keyword{internal}
