% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{.besd_resolve_palette}
\alias{.besd_resolve_palette}
\title{Built-in palette resolver}
\usage{
.besd_resolve_palette(palette = c("imms", "okabe_ito"), n, levels = NULL)
}
\description{
Built-in palette resolver
}
