% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\docType{data}
\name{.BESD_SEP}
\alias{.BESD_SEP}
\title{Unit Separator: safe delimiter for packed multichoice strings}
\format{
An object of class \code{character} of length 1.
}
\usage{
.BESD_SEP
}
\description{
Unit Separator: safe delimiter for packed multichoice strings
}
\keyword{datasets}
