% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{.strip_non_alpnum}
\alias{.strip_non_alpnum}
\title{"Normalise" a string by converting to alpha-numeric}
\usage{
.strip_non_alpnum(x)
}
\description{
"Normalise" a string by converting to alpha-numeric
}
