% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{modify_dictionary}
\alias{modify_dictionary}
\title{Add / remove items from dictionary}
\usage{
modify_dictionary(
  df = NULL,
  id_out = NULL,
  dict = besd_dictionary(),
  replace = TRUE
)
}
\arguments{
\item{df}{Data frame/tibble of new/updated items. Must include dictionary columns
including \code{item_id}.}

\item{id_out}{Character vector of item_id values to remove.}

\item{dict}{Existing dictionary.}

\item{replace}{If TRUE, overwrite existing rows with same item_id.}
}
\description{
Items being added must include \code{item_id} and all required columns.
}
