% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-besd.R
\name{plot_besd_spider}
\alias{plot_besd_spider}
\title{Spider (radar) plot for BeSD top-box by country}
\usage{
plot_besd_spider(
  sum_tbl,
  compare = "all",
  focal_country = NULL,
  topbox_levels = NULL,
  item_ids = NULL,
  palette = NULL,
  base_size = 12,
  wrap_width = 22,
  ncol = 3,
  spider_scale = 0.8,
  label_padding = 12,
  outer_padding = 12,
  facet_padding = 1.2,
  overlay = FALSE
)
}
\description{
Spider (radar) plot for BeSD top-box by country
}
