% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbesd-package.R
\docType{package}
\name{rbesd-package}
\alias{rbesd}
\alias{rbesd-package}
\title{rbesd: WHO/UNICEF BeSD Quantitative Toolkit Analysis}
\description{
The \code{rbesd} package provides an end-to-end workflow for WHO/UNICEF Behavioural and
Social Drivers (BeSD) quantitative toolkit data. It includes:
\itemize{
\item Standardisation and validation against a common item dictionary.
\item Storage of data and metadata in a \code{besd_data} object.
\item Summaries, scoring, and plotting utilities.
\item Regression wrappers for binary and ordinal outcomes.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexdefig/rbesd}
  \item Report bugs at \url{https://github.com/alexdefig/rbesd/issues}
}

}
\author{
\strong{Maintainer}: Alexandre de Figueiredo \email{alex.defigueiredo@lshtm.ac.uk}

}
\keyword{internal}
