% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besd-class.R
\name{besd_info}
\alias{besd_info}
\title{Access BeSD metadata}
\usage{
besd_info(x)
}
\arguments{
\item{x}{A \code{besd_data} object.}
}
\value{
A named list with entries \code{besd_dict}, \code{dem_dict}, \code{besd_items},
\code{dem_items}, \code{country_col}, \code{stratum_col}, \code{psu_col}, \code{weight_col},
\code{id_col}, \code{mapping}, and \code{meta}.
}
\description{
Returns dictionaries, column names, and metadata stored as attributes on a
\code{besd_data} object.
}
