% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-besd.R
\name{besd_topbox}
\alias{besd_topbox}
\title{Build a top-box tibble from a BeSD summary tibble}
\usage{
besd_topbox(
  sum_tbl,
  topbox_levels = NULL,
  prefer_regex = c("^Yes$", "Strongly agree|Completely agree", "Agree a lot|Agree",
    "Very likely|Definitely", "A lot|High"),
  include_item_types = c("binary", "ordinal", "categorical", "unknown")
)
}
\description{
Build a top-box tibble from a BeSD summary tibble
}
