% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.besd_apply_response_order}
\alias{.besd_apply_response_order}
\title{Apply dictionary-based response ordering}
\usage{
.besd_apply_response_order(sum_tbl)
}
\description{
Apply dictionary-based response ordering
}
\keyword{internal}
