% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-besd.R
\name{plot_besd_bars}
\alias{plot_besd_bars}
\title{Bar plots for BeSD response distributions}
\usage{
plot_besd_bars(
  sum_tbl,
  include_item_types = c("binary", "ordinal", "categorical", "unknown"),
  include_multichoice = FALSE,
  sort_bars = FALSE,
  palette = NULL,
  base_size = 12,
  label_pct = TRUE,
  label_min = 6,
  wrap_width = 50,
  combine_domains = FALSE
)
}
\description{
Bar plots for BeSD response distributions
}
