% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_besd_demographics_table}
\alias{plot_besd_demographics_table}
\title{Plot a demographic summary as a "pretty table"}
\usage{
plot_besd_demographics_table(
  dem_tbl,
  country,
  max_levels = 12,
  sort_by_pct = TRUE,
  fill_by = c("pct", "none"),
  fill_high = "#4F8D9A",
  base_size = 12,
  wrap_width = 30
)
}
\description{
Plot a demographic summary as a "pretty table"
}
