% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_besd_multichoice_bars}
\alias{plot_besd_multichoice_bars}
\title{Bar plots for multichoice summary items}
\usage{
plot_besd_multichoice_bars(
  sum_tbl,
  item_ids = NULL,
  top_n = 10,
  palette = NULL,
  base_size = 12,
  wrap_width = 32
)
}
\description{
Bar plots for multichoice summary items
}
